[Refactor] #67 Organizing folder structure
