+++ campair/campair/Application/Images/Equipment/경량_체어.png
