+++ campair/campair/Model/DictionaryRepository.swift
@@ -1,0 +1,8 @@
+//
+//  DictionaryRepository.swift
+//  campair
+//
+//  Created by Lee Myeonghwan on 2022/06/17.
+//
+
+import Foundation
