--- campair/campair/Utility/UtilityPlaceHolder.swift
@@ -1,8 +1,0 @@
-//
-//  UtilityPlaceHolder.swift
-//  campair
-//
-//  Created by Shin Jae Ung on 2022/06/10.
-//
-
-import Foundation
