+++ campair/campair/View/Dictionary/Content/DictionaryContentView.swift
@@ -1,0 +1,20 @@
+//
+//  DictionaryContentView.swift
+//  campair
+//
+//  Created by Lee Myeonghwan on 2022/06/17.
+//
+
+import SwiftUI
+
+struct DictionaryContentView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct DictionaryContentView_Previews: PreviewProvider {
+    static var previews: some View {
+        DictionaryContentView()
+    }
+}
