+++ campair/campair/ViewModel/Dictionary/DictionaryContentViewModel.swift
@@ -1,0 +1,8 @@
+//
+//  DictionaryContentView.swift
+//  campair
+//
+//  Created by Lee Myeonghwan on 2022/06/17.
+//
+
+import Foundation
