+++ campair/campair/ViewModel/Dictionary/DictionaryDetailViewModel.swift
@@ -1,0 +1,8 @@
+//
+//  DictionaryDetailViewModel.swift
+//  campair
+//
+//  Created by Lee Myeonghwan on 2022/06/17.
+//
+
+import Foundation
