+++ campair/campair/ViewModel/Dictionary/DictionaryMainViewModel.swift
@@ -1,0 +1,8 @@
+//
+//  DictionaryMainViewModel.swift
+//  campair
+//
+//  Created by Lee Myeonghwan on 2022/06/17.
+//
+
+import Foundation
